Add PedagioDAO class for MongoDB interaction
Introduced a new `PedagioDAO` class in the `Kria.Core.Pleno.Lib.DAO` namespace to interact with the "Pedagios" MongoDB collection. This class implements the `IPedagioDAO` interface and includes the `ObterTodos` method to retrieve all documents as a list of `Pedagio` objects. Added necessary `using` directives for MongoDB and other functionalities. Defined the `IPedagioDAO` interface, though it currently lacks method declarations.